Legistar API Methods Reference

A data frame with reference information from the Legistar Web API documentation: https://webapi.legistar.com/. Organized in a Google Sheet for use with this package: https://docs.google.com/spreadsheets/d/1Py-Oqyb8kW13x0_BjGJS-fw1EdZzqBeIsjxvQvBV5dM/edit?usp=sharing Last updated 2024-05-12.

Usage

legistar_methods

Format

A data frame with 124 rows and 8 variables:

type

API method type

method

HTTP Request Method

template

Template string

url

Help URL

description

API method description

response

Expected API method response

modelName

Corresponding modelName for response

nm

Snake case method name

Source

https://webapi.legistar.com/